+++ src/scope/network/check-version-compatibility.js
@@ -15,13 +15,33 @@ const createMinorMessage = (remoteVersion, currentVersion) =>
   );
 
 export default function checkVersionCompatibility(remoteVersion: string) {
-  if (semver.major(remoteVersion) > semver.major(BIT_VERSION)) {
+  // In case the client is newer than the server version don't check computability
+  // (Should be change in the future, but right now we won't release a client which we don't support
+  //  on the server)
+  if (semver.gte(BIT_VERSION, remoteVersion)) {
+    return;
+  }
+
+  const remoteMajor = semver.major(remoteVersion);
+  const remoteMinor = semver.minor(remoteVersion);
+  const remotePatch = semver.patch(remoteVersion);
+  const localMajor = semver.major(BIT_VERSION);
+  const localMinor = semver.minor(BIT_VERSION);
+  const localPatch = semver.patch(BIT_VERSION);
+
+  if (remoteMajor > localMajor) {
     loader.off();
     console.log(createMajorMessage(remoteVersion, BIT_VERSION)); // eslint-disable-line
     return;
   }
 
-  if (semver.minor(remoteVersion) > semver.minor(BIT_VERSION)) {
+  if (remoteMinor > localMinor) {
+    loader.off();
+    console.log(createMajorMessage(remoteVersion, BIT_VERSION)); // eslint-disable-line
+    return;
+  }
+
+  if (remotePatch > localPatch) {
     loader.off();
     console.log(createMinorMessage(remoteVersion, BIT_VERSION)); // eslint-disable-line
   }