+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- improve checkVersionCompatibility between server and client
 ## [0.10.9-dev.1] - 2017-10-02
 
 - handle invalid bit.json